tests/wpt/tests/css/css-shapes/parsing/shape-outside-shape-valid.html

<!DOCTYPE html>
<title>CSS Shapes Module Level 2: parsing shape-outside with valid shape() values</title>
<link rel="author" title="Jason Leo" href="mailto:cgqaq@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-shapes-2/#shape-function">
<link rel="help" href="https://www.w3.org/TR/css-shapes-1/#shape-outside-property">
<meta name="assert" content="shape-outside supports the shape() function.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
test_valid_value("shape-outside", "shape(from 20px 40px, line to 20px 30px)");
test_valid_value("shape-outside", "shape(from 0 0, line to 100% 100%)", "shape(from 0px 0px, line to 100% 100%)");
test_valid_value("shape-outside", "shape(from 20px 40px, move to 20px 30px, line by 20px 30px)");
test_valid_value("shape-outside", "shape(from 20px 40px, move to 20px 30px, hline to 100px)");
test_valid_value("shape-outside", "shape(from 20px 40px, move to 20px 30px, vline to 100px)");
test_valid_value("shape-outside", "shape(from 20px 40px, curve to 20px 20px with 10px 30px)");
test_valid_value("shape-outside", "shape(from 20px 40px, curve to 10% 20% with 10px 30px / 12px 32px)");
test_valid_value("shape-outside", "shape(from 20px 40px, smooth to center 20%)");
test_valid_value("shape-outside", "shape(from 20px 40px, arc by 20px 20px of 10%)");
test_valid_value("shape-outside", "shape(from 20px 40px, arc by 20px 20px of 10% 20% cw large)");
test_valid_value("shape-outside", "shape(from 20px 40px, arc by 20px 20px of 10% 20% cw rotate 12deg)");
test_valid_value("shape-outside", "shape(from 20px 40px, line to 100px 40px, line to 100px 100px, close)");
</script>
